+++ experimental/behave/src/main/kotlin/net/corda/behave/node/Distribution.kt
@@ -95,7 +95,7 @@ class Distribution private constructor(
 
         private val distributions = mutableListOf<Distribution>()
 
-        private val nodePrefix = stagingRoot / "deps/corda"
+        private val nodePrefix = stagingRoot / "corda"
 
         val MASTER = fromJarFile("corda-master")