+++ experimental/behave/src/scenario/kotlin/net/corda/behave/scenarios/steps/VaultSteps.kt
@@ -21,6 +21,7 @@ class VaultSteps : StepsBlock {
 
         Then<String, Int, String>("^node (\\w+) vault contains (\\d+) (\\w+) states$") { node, count, contractType ->
             try {
+                @Suppress("UNCHECKED_CAST")
                 val contractStateTypeClass = Class.forName(contractType) as Class<ContractState>
                 if (vault.query(node, contractStateTypeClass).size == count)
                     succeed()
@@ -33,7 +34,7 @@ class VaultSteps : StepsBlock {
 
         Then<String, Long, String>("^node (\\w+) vault contains total cash of (\\d+) (\\w+)$") { node, total, currency ->
             val cashStates = vault.query(node, Cash.State::class.java)
-            val sumCashStates = cashStates.filter { it.state.data.amount.token.product.currencyCode == currency }?.sumByLong { it.state.data.amount.quantity }
+            val sumCashStates = cashStates.filter { it.state.data.amount.token.product.currencyCode == currency }.sumByLong { it.state.data.amount.quantity }
             print((sumCashStates))
             if (sumCashStates == total)
                 succeed()